Revert "merge"
This reverts commit f1518de35e14954060a031b363ab344035a8773d, reversing
changes made to 2a162c91b257018f0586dc08f076a0e424ac6b10.

+++ Models/Loyalty.cs
@@ -15,7 +15,7 @@ public partial class Loyalty
 
     public int? TicketCount { get; set; }
 
-    public decimal? PriceMultiplier { get; set; }
+    public int? PriceMultiplier { get; set; }
 
     //Relationships
     [ForeignKey("EventUserId")]
